***CONTAINS FIXES THAT NEED BACKPORTING TO EARLIER PATCHES IN MY STACK***
add a file:// staging mechanism which stores files inside a wrapped function

this allows small local file:// files to be staged-in to the remote side using only the existing
function call communication channels

this is more intended as a proof of concept than for real world use, particularly
for testing out the semantics of local file:// files no longer being "hands off"
or "unstaged"

also adds in a new staging API call that wasn't needed for previously-implemented methods

I'm pretty concerned about how error handling happens here with stacked unwrappers:
we might have a RemoteException wrapper on the very outside, or a RemoteException wrapper
on the very inside, but not in-between (assuming that staging code just throws an
exception and that it is wrapped in a RemoteExceptionWrapper at the executor level)

It's pretty awkward to call the unwrap function - the value to unwrap doesn't appear until
inside the AppFuture callback... ongoing (#722-related) refactoring of this should make
this code path easier. It's not massively morally different to swapping out a
RemoteExceptionWrapper for a real exception, for example.


This code also temporarily removes wrap_error - I will let the outer executor catch
errors, and this way lets exceptions abort stage out -- which doesn't happen otherwise...

This code also encounters issue #1197 in the test suite, because it is the first
provider to set local_path for file: scheme files