Move self.statefile to make RepresentationMixin happier
Previous to this commit: if the boto client initialization
throws an exception, then the exception handling of that
throws another exception from __repr__() from
RepresentationMixin, as it cannot list out all of the
initializer fields.

This commit moves self.statefile to earlier in the
initializer, before the above boto client initialization.