Add a site test for ec2 (#1136)
This requires user_opts to have been suitably configured, so will not run in CI.

Awkwardnesses:

i) template.py needs to be adjusted to install the same version of parsl as is being submitted from, rather than whatever the most recent release is - as needs to happen with any non-trivial development version of parsl submitting to EC2.

ii) This will not run inside pytest - as with a bunch of other tests, some apps run inside pytest cause more modules to be loaded than run outside pytest, the reason for which I don't understand. In this case, that means the workers try to import the test script, which causes several problems - at least, that PARSL_TESTING is not installed, that the test-requirements dependency stack is not installed, that user_opts is not set correctly on the remote side.