Add control of logging level to the Config, stream_ and file_logger
This adds new settings in the `Config` class which enables setting
the logging levels to the stream_logger and the file_logger.

The main use case here is for those directly invoking the
`DataFlowKernel` with a `Config` who don't want the output loggers to
be automatically set to `DEBUG` level.

Default behavior should be identical to the current behavior and this
*should* be an extension of the existing API without breaking backwards
compatibility.

It is a rather simplistic change, so I'm not sure I made the changes
everywhere that is needed nor have I figured out how to properly test
this, so any comments or suggestions would be helpful.

Change is related to a conversation @yadudoc and myself had previously.