Add Workqueue executor (#771)
This is the initial barebones implementation of the WorkQueue executor.

The general idea is that it creates a shared task dictionary to relate Parsl tasks to their future, and a queue to send tasks to the worker.

At executor start, a thread is started to handle the WorkQueue operations, where the executor sends tasks through the queue to this thread. The python executables are pickled to a file, and this file is sent along with the task to the worker to be executed with the worker script (work_queue_worker.py) that is sent along with each task.

The worker thread monitors the queue, and when it takes values from the queue it submits them, and then waits with a threshold of five seconds for a task before repeating. When a task is completed, it unpickles the result, and updates the futures that were stored in the dictionary previously.