not all exceptions can be serialised by remoteexceptionwrapper
for example, GlobusConnectionError cannot

when this has happened with parsl exception classes, those classes
have been modified so they are serialisale (I think? PRs?)

but that approach doesn't work so easily with non-parsl exceptions

the specific case I am encountering is:

Traceback (most recent call last):
  File "/home/benc/parsl/src/parsl/parsl/dataflow/dflow.py", line 254, in handle_exec_update
    res.reraise()
  File "/home/benc/parsl/src/parsl/parsl/app/errors.py", line 160, in reraise
    v = dill.loads(self.e_value)
  File "/home/benc/parsl/virtualenv/lib/python3.5/site-packages/dill/_dill.py", line 317, in loads
    return load(file, ignore)
  File "/home/benc/parsl/virtualenv/lib/python3.5/site-packages/dill/_dill.py", line 305, in load
    obj = pik.load()
TypeError: __init__() missing 1 required positional argument: 'exc'
dflow.py                   278 INFO     Task 1 failed after 0 retry attempts
errors.py                  158 DEBUG    Reraising exception of type <class 'globus_sdk.exc.GlobusConnectionError'>

which is a problem serialising GlobusConnectionError.

This shouldn't need to happen - because globus stuff happens in the local python process - so no
serialisation is needed.

But RemoteExceptionWrapper is too eager in its pickling.

Is it possible to push the serialisation of constituents into a method invoked only
on serialisation and deserialisation, making that serialisation lazier, and so not
happening if we keep the RemoteExceptionWrapper within the same process?

Exceptions which cannot be serialised will still not work with this wrapper if an
attempt is made to move them between processes.

But exceptions which are wrapped but not moved between processes will make it through
the wrapper ok.

ALTHOUGH

the weird thing here is that if i just remove all the serialisation, everything pickles up ok
with the htex executor. which is nice... maybe the presence of Traceback has fixed the need to pickle
explicitly? maybe htex uses a different pickler to IPP? in which case I should test with ipp too?