Refactor _gather_all_deps
This removes some repeated code. In a future PR,
that code will be a bit more complicated so it
is useful to have it not repeated.

Because an inner function is used, an integer
cannot be used as a count any more as it cannot
be modified from that inner function.

Instead, a new list, which is a mutable structure,
is used to accumulate the count instead.