Treat DataFuture inputs as dependencies
Prior to this PR, DataFutures were treated, dependency-wise, like
any other input value, and did not cause any deferral of execution.

This meant that output DataFutures from one app passed as inputs
to a second app would not cause the second app to wait for the
first app to complete.

Fixes issue #1125