Give Globus stageout code the chance to set File.localpath before main task
This makes Globus stageout localpath behaviour the same as stagein behaviour:
if staging will be happening to some non-default working directory, then
localpath will be set to point to that actual location of staging.

Because stageout completion is not tested at the moment, the behavioural change is:

Before this PR, if user_opts['globus']['path'] is not equal to the current working directory:
globus stageout test would complete with a successful exit code before
stageout happens (see issue #778); but a transfer job would be left in globus
stalled with a file-not-found error (forever, or at least for a day) in some
circumstances.

After this PR: globus stageout test will complete with a successful exit code before
stageout happens; a globus transfer job will be made and this will successfully
stage out the file.

Fixes issue #1173