Make explicit copies of File objects per execution
This addresses issue #1197

Prior to this PR, File objects were shared or not-shared between
different pieces of parsl code based on the nature of the
executor.

For example, modifying a File object inside a stage-in task
running in the local thread executor would modify the original
File object passed into the app (because the thread executor passes
around the original object), but modifying a File object inside
a stage-in task running in the htex executor would not modify
the original File object, as File objects are copied (implicitly
by serialization of app arguments).

Also, the local_path attribute which was sometimes set to provide
an execution-side filesystem path did not always propagate properly.
This was generally disguised by the presence of a vestigial
enumeration of http/ftp/globus file schemes which should have been
removed as part of the earlier staging API work.

Also, sometimes when the local_path attribute was set, that
sometimes propagated too much: if staging ended up modifying the
original source File object (as described above) then this would
modify the File object even when passed to a different app
invocation - most badly, an app invocation on another site where the
execution-side filesystem path may be different to the one
already set.

This PR more explicitly copies File objects at the point that they
are used for an app invocation; staging code can modify only that
invocation-specific copy, leaving the original pristine.

This PR removes the above-mentioned vestigial scheme enumeration;
it is now an error to ask for the local filesystem path of a File
which does not have local_path set, except in the case of a file:
schemed file. It is now a requirement that all staging providers
set File.local_path.

This PR changes the way in which the existing staging providers
set File.local_path so that it is always set properly.