Remove vestigial scheme enumeration
It should be up to staging providers to specify the appropriate
local_path for all schemes, except for file: which has an
implicit meaning on the submit side POSIX filesystem.

File.filepath will now throw an exception in many cases where
this does not make sense. As a consequence, some log messages
throughout the codebase are adjusted to not call filepath,
and instead use the full File repr.

Consequent changes are needed for the existing staging provider
to make them work.

Staging providers that attempted to implement a differently
named scheme would fail.