Make pipeline test assert that result files don't have remote path set
This is to catch issue #1197, which I have experienced in real life
when developing a post-#1201 staging provider - staging providers might
modify the submit-side File object in such a way that attempts to pull
the submit-side posix path out of that File object instead give the remote
path.

This is likely to only be a problem when implementing interesting
file:// scheme providers - this test (correctly) fails for me on
one such provider that I have been working on.