Add a string deprecation warning when DataFutures are constructed with strings
This manifests in the test suite when app outputs are specified with strings
rather than File objects; and for those strings, File objects are constructed.

However:
 * these Files are not staged out properly according to staging rules;
 * the behaviour of strings on inputs is to be passed through unmodified to
   apps, rather than being wrapped as files.

The longer term behaviour should be to remove this special casing, but it will
break user code that relies on specifying local unstaged files as strings.