Give staging providers copies of `File` rather than the originals.
Addresses #1197 - where File.local_path is sometimes too global,
when that state should properly be constrained to a single execution
site or to the submit side.

I suspect this happens when using globus staging to two sites
(where the local_path is set in the submit side python process),
and I have seen it happen when developing staging providers that
implement the file: scheme in the single site case; because
File.filepath can return the remote side file path instead of
the submit side filepath.