Remove conftest imports that interfere with pytest fixtures
Importing conftest into a test causes pytest to see fixtures as
new fixtures. This means if trying to scope the load_dfk as
session, rather than module, it tries to start that fixture again,
rather than continuing to use the existing DFK.

This at least causes the test to fail, because the newly
imported fixture attempts to load a DFK when one is already loaded;
and secondly attempts to load a DFK which is potentially very
expensive in a cloud/HPC environment.

These tests were using load_dfk to do pytest-like config loading
without using pytest, in __main__. The correct way to do this
should be to use pytest, and this PR removes those non-pytest
__main__ sections.