Add whole-run DFK pytest fixture. Remove many-config capability.
This change will (in non-local mode) load a single DFK for a specified
config, which will be used for all tests in the run.

For local mode, new ways of initialising tests are added: local_config,
local_setup and local_teardown, that can be used to write 'local'-mode
tests which do not interfere with other tests - the parsl test suite
local mode tests, at present, do lots of global stuff, and these new
parameters and methods allow that globus stuff to be more constrained
to individual test runtime.

This conflicts with the implementation of multiple test configurations
in one run - that feature causes the DFK to be loaded/unloaded for each
test module as pytest sweeps over the list of configs, even if the list
only has one entry.

Multiple test configuration handling is removed, and --configs is renamed
to --config.

The behaviour can be demonstrated by running:

pytest parsl/tests/test_bash_apps/test_apptimeout.py \
  parsl/tests/test_bash_apps/test_basic.py \
  --config parsl/tests/configs/local_threads.py --setup-show

before and after, where --setup-show will show when each fixture is executed