Move extreme scale executor mpi4py import requirement into start
Prior to this commit, attempting to construct a Config involving
the extreme scale executor would fail if mpi4py was not installed
with an exception message informing the user about that requirement.

After this commit, that exception will be thrown later, at the point
that a user attempt to load a DFK with that config.

This allows unused extreme scale executor configs to be imported
safely - most immediately, for test rearrangements in PR #1210