Remove use of 'noci' pytest marker (#1230)
This leaves the marker implementation in case it is needed in the
future, but removes its use in favor of implementing a
`pytest_ignore_collect` function in `conftest.py`. The advantage of this
is it removes the need to append each pytest invocation with `-k not
noci`, which is needed by definition. The disadvantage is that, since we
mix pytest and non-pytest tests in the testing directory, this
obfuscates which tests will be ignored by pytest.