Tidy cluster provider 'abstract base class' behaviour
ExecutionProvider has submit and cancel methods which are labelled
as @abstractmethods.  Prior to this PR, ClusterProvider implemented
those with stubs that always failedd; this removes some of the
benefit of using @abstractmethods. This PR removes those stubs, so
that ABC behaviour will still happen for those methods.

ClusterProvider introduces a new method which must be overridden
by subclasses, _submit. As with submit and cancel, this was implemented
as a non-functioning stub. This PR replaces that with an
@abstractmethod.