Make multinode tests not import config by default
Previously:
These tests import a config file which uses an SSH channel to
talk to a remote host; that config file establishes the SSH
channel at the point at which the configuration is imported,
if the appropriate cori user opts are set. This means that an
SSH attempt is made even when running with (for example) the
local threads config file, and the test run fails if that SSH
attempt fails.

After this PR:
These tests will only import the config file when run in 'local'
config mode.