Added WQ Executor Functionality to Increase Throughput (#1152)
This pull request makes modifications to the WorkQueue Executor to provide better throughput. Instead of using the ipyparallel.serialize library to serialize Python functions, a switch has been implemented for the Executor that allows the user to choose whether to transport the function as source code. This greatly diminishes the overhead for sending and receiving task information that was due to the expensive serialization. It also uses the tblib library to serialize traceback objects for tasks that failed.

To test full functionality, you can use the "source" option in the parsl/tests/workqueue_tests/wqex_local.py to switch between transferring function data as a serialized object or through source code.