Fix race condition in checkpoint test (issue #681)
The test previously made the incorrect assumption that
future.result() returning meant that the checkpoint for
the corresponding app had been written out.

The assumption made by the test after this commit is:

If future.result() has returned, and then *after that*
dfk.cleanup() has been invoked and returned, then
the checkpoint for the corresponding app has been
written out.