Fix depfail handling introduced by PR #1200, issue #1257
In the codebase, dependency failures are handled by creating an executor
future that is populatd with an exception immediately rather than waiting
for some execution to happen.

Both before and after PR #1200, the callback setup differs between
executor futures that come from an actual executor vs for executor
futures that contain a dependency failure.

Prior to #1200, an regular executor future would get two callbacks
registered: one via AppFuture.update_parent to inform the AppFuture
and one to inform the DFK of the result. However, a DependencyFailure
executor future did not have both of those registered - it only had
the AppFuture.update_parent callback registered.

would now happen via the DFK callback; in the dependency failure case
this was not true (just as it had not been true before PR #1200).

This commit makes the handling of regular executor futures vs
dependency failure executor futures more consistent, by reducing the
amount of code in the no-failure vs failure paths and using the
same callback registration code no matter which of these two ways the
executor future is created.