Remove ability for providers to be not-scalable
This code path is not used, and providers have a more nuanced
ability to control their own scaling, because they are allowed
to not launch a block when asked.

This does not remove the ability for executors (rather
than providers) to be marked as scalable or not. Executors which
previously sometimes inherited their scalability from their
provider now also report themselves as scalable when given a
provider.

Fixes #1304