Tighten pipeline tests to assert more.
In some cases when developing staging providers, the pipeline
test fails due to a buggy provider, but giving an error that is
not immediately clear - rather than indicating that an input
file is missing, the test instead starts counting from 0 again.

This patch adds in an explicit test for expected input files to
aid in detecting that particular failure mode.

This patch also adds checks to ensure that returned outputs are
DataFutures - this catches the case that file stage-out code is
incorrectly returning a File rather than DataFuture for that
File.