Make DFK optionally not initialize any logging
The use case for this is when parsl is used as a library
as part of a bigger system which wants to configure logging
in a way that makes sense for that bigger system as a whole.

In such a situation, parsl should step out of the way
and let that system do what it wants.