Add test mark for tests affected by #363
Tests which inspect stdout/stderr cannot always expect
to find the stdout/stderr files on the submitting file
system. That is issue #363

This commit adds @pytest.mark.issue363, so that
pytest can be invoked with parameter:

   -k 'not issue363'

which skips tests affected in that way.

I have tested this with the AWS provider (as part of a
set of various other AWS test fixes and local configs).