Tighten KeyError handling that was falsely skipping failing tests (#1382)
This pulls out all KeyError handling except at the collect stage.

At the collect stage, keyerrors only cause a skip if the top of the stack
trace is in parsl/tests/configs/ - so you can only reference user_opts there,
not anywhere else.

KeyErrors happening outside of the collect stage, even if they are in a
config file (for example, if a test local_setup() does an import of a
config), will result in test failures.

Broadly this means that KeyErrors happening in parsl/tests/configs/*
imported at the top level into a test will cause a skip; otherwise KeyErrors
will be treated as errors.