Remove Hub->MonitoringHub stop_q
Prior to this commit, the Hub instance sent a message on
the stop_q queue to indicate that it has performed all of
its shutdown, before immediately ending.

The MonitoringHub close method which triggers Hub shutdown
waits to receive this message before performing a terminate
on the already-terminating Hub.

This can cause a deadlock if the Hub process has already
died for some other reason: the MonitoringHub will wait for
a stop message that will never arrive.

This mechanism appears to be more complicated than
necessary: the MonitoringHub can signal that the Hub
process should stop, and then wait for the Hub process to
end using the Process API.

This commit removes the whole stop_q mechanism, hopefully
removing this potential deadlock.