Inline AppFuture parent_callback into DFK
This merges into one place two code paths that work together
to deal with executor future completion, and which are probably
better understood when placed next to each other.

This commit should only inline the call so should not change
functionality.

This is the final piece to merge of PR #722