Fix bash_app walltime test (#1433)
Prior to this commit, this test did not check that timeout
behaviour actually happened. After this commit, the test
checks for an AppTimeout, in the style of the python_app
test in parsl/tests/test_error_handling/test_python_walltime.py

This commit also adds a test that apps will pass without
an exception if they run within the specified walltime.