Fix status checking in LocalProvider non-LocalChannel path
Prior to this commit, the pid to inspect was not properly
passed in, and the status check always returned that a
remote process was not running.

When strategy was enabled, this meant that the strategy
would keep launching blocks, even when they were starting
and running successfully.

Part of plan in issue #1448