3. Replace job states with an explicit enum
FOR ME: check difference between CANCELLED and CANCELED - I've maybe
typoed that a bit?

CHANGE REQUEST: There is a removed if: in the CobaltProvider that has an
open question right next to it - that question probably should be
resolved before this is merged.

CHANGE REQUEST: at least ad_hoc.py refers to "status strings"
which are not a thing any more
        list of status strings [JobStatus]


This doesn't update the job states used in LocalProvider, but PR #1454
removed that table and should merge with this pretty cleanly.

Likewise PR #1465 removes an apparently unused translation table in the
pbspro provider