Make LocalChannel execute_wait set pgid like execute_no_wait
execute_no_wait sets the process group of launched commands so that
local and ad-hoc providers can kill those launched commands later by
killing the whole process group.

Before this commit, execute_wait did not do that, so execute_wait could
not be used in the same way - instead, processes were left in the same
process group as the parent, and local/adhoc providers kill the entire
parent process group (for example, the entire containing workflow).

This commit also amends the documentation for the Channel base class to
clarify this requirement.

This is part of the plan in issue #1448