Make launcher scripts fail on error. This will only be useful if/when launcher failures are reported back to the user. (#1466)
There are two types of errors that this will catch: errors in the assumptions of the launcher script (such as filesystem errors while writing sub-scripts or missing mpirun/srun/parallel/etc.) and application/block errors if the launching command propagates them. Both are useful to be reported. Moreover, from the school of thought of "all errors should be appropriately handled or reported", this is a must.