Fix jumbled legends and colors in workflow dags.
Prior to this, colours and labels for jobs in the two DAG graphs were
jumbled up.

The code incorrectly assumed that iterating over a dict would iterate in
the order that the entries appear in a hard-coded literal dict in the
source tree.

For example:

The code assumed that the first element added to the node_traces list, and
thus the element referred to by node_traces[0], would be the element in
groups_list with value (0,...)

This incorrect assumption led to adding DAG nodes to the wrong Scatter
object.

This patch changes the way that the node_traces list is constructed by
populating it with marker places and then using index notation to set the
appropriate entry in the list based on hard-coded index.

Further, that wrong scatter object then had its colour mutated. This patch
sets the colour correctly on initial creation and does not change it again.

For the DAG coloured by state plot, the naming if-statement had no final
else: case, leading to some nodes being incorrectly labelled. This patch
adds an 'Unknown' case which is reached, for example, in jobs which are
never launched due to dependency failure.