Make config-changing staging test run in local mode
Prior to this, one of the staging tests added an executor to
the DFK, to check that staging still worked in that mutated
environment.

However, this test leaves the mutated environment in place,
meaning that subsequent tests will now sometimes run on that
added executor instead of on the command-line specified
executor.

This commit changes that test to run in local mode, so that
there is a DFK specific to just this test module for the
additional executor test.