Infer submitted and running times in gantt chart (#1525)
Prior to this commit, the gantt chart appeared to default to the year 2000 for some of the times
used. This mean that the default scale of the gantt chart was useless (decades, rather than
seconds) for a CI test run.

After this commit, if either the submitted or running times are None, then the return time is used.