Fix tasks_per_app plot enough to visualise a CI run. (#1524)
Before this patch, the tasks_per_app plot could not be displayed when
viewing a CI test run monitoring.db. This patch fixes the plot so
that a CI run can be visualized.

i) the plot was expecting ints, but was getting numpy.float64
from the underlying database.

ii) the plot was expecting ints, but the time_running fields are
None/NaN because checkpointed tasks never start running.