Make memoization hashing pluggable by type
By default, arbitrary types cannot be used as memoization keys. This commit
adds a test case for that.

Basic python types are hashed as before.

Dicts and lists are now normalized more usefully.

kwargs now normalise more usefully. This commit adds a test
case for that.

This fixes issue #235

TODO:
if we're allowing user-registered id_for_memo calls, then
the invocations of id_for_memo in serialisation need to be
more resilient to bugs in user code. at the time of writing,
an exception thrown in a user registered id_for_memo sometimes
seems to hang parsl.
(this could be testable with a test-suite provided error
throwing test?)