Prohibit ERROR log messages in pytest, with an escape clause
This introduces a couple of fixtures, a log handler and a context manager
which together check that log messages of severity ERROR or higher are
not logged during test runs, except during periods when a test is in a
permit_severe_log context.

If an ERROR is logged during a test, that test will fail.

TODO: move all this into its own source file

TODO: tidy naming

TODO: what about logs that happen when DFK is initialised in
other fixtures, or at shutdown after the last test?