Make IPP executor return scaled out blocks and not log a spurious error
Indentation meant that an else clause was attached to the scaling for
loop, and so ran at the end of the loop every time, rather than to the
if statement checking if a provider is available.

I discovered this as part of #1541, auditing ERROR log messages in the
test suite.