Harden error handling around address_by_query
Prior to this, the HTTP call sometimes returned formatted HTML rather
than a plain IP address, in certain error modes. This then propagated
without warning deep into the rest of parsl, causing an obscure error
later on.

After this:

Exception handling is made tighter around each resolution method.

address_by_query will raise an exception if it receives an HTTP
response other than status code 200 OK.

This attempts to address #1456 but it is hard to test that because
I'm not sure what ipify is returning for a status code in this
particular error case.