Adding a new "force" option to the executor's scale_in interface.
A force option allows us to indicate whether we want to force kill
regardless of whether blocks are idle. We also now sort blocks by
the number of tasks outstanding on it, so as to attempt killing off
the least loaded ones. When force=False we only pick blocks that are
idle.