Add a facility to label some arguments in an app invocation as not memoized
This means that the values will not be used for making equality comparisons between two different
invocations - a difference in those values will not cause the equality to be made unequal.

The way this is implemented, turning on/off this mode for a particular parameter *will* cause the
equality to be unequal - because with the mode on, the argument value will not go into the
checkpoint hash, but with the mode off, the value will go into the checkpoint hash.

This makes sense if understanding checkpoint equality as being based on a hash - but needs
explaining because if the mental model of this is "compare the parameters but ignore the ones that
are listed as ignorable" this isn't so clear.

TODO: update checkpointing docs

TODO: merge in `test` patch from prior to this one

TODO: test with autonamed stdout/stderr