Make IPyParallelExecutor a ParslExecutor again, so remaining site tests typecheck
Some site-specific tests are left for future users to port to HTEX. These tests
still need to typecheck, and so IPyParallelExecutor still needs to be a ParslExecutor.

As a consequence of that, the "ipp is gone" exception message is now issued in
__new__ rather than init, so that it happens before the abstract base class code
discovers that IPyParallelExecutor doesn't actually implement all the methods
needed by ParslExecutor.