Handle the case where a task does not have a hashsum for memoization
This case was occuring when a task set for memoization fails due to a
dependency failure. In this situation, the inputs of the task have not
been hashed and stored in hashsum (and do not even exist because the
task that produces them failed).

Previously, a key error was thrown, propagating up to handle_app_update.
Now, the memo table is silently not updated.