Make monitoring workflow name work with recent typeguard and ignore __init__.py
Recent versions of typeguard place typeguard/__init__.py in the stack instead
of typeguard.py, so the code which skipped typeguard.py in workflow name
detection didn't skip typeguard and was picking __init__.py as the workflow
name.

Probably in general __init__.py is never the right name for a workflow in
monitoring - so this PR skips that name in all circumstances.

If the code runs out of names to try, then the code now defaults to a
specific value rather than leaving the workflow name un-set.