Allow task_hashsum to be null in monitoring
This was introduced to monitoring in PR #1584 but the
schema required it to be NOT NULL.

This field is NULL when a task fails before its hashsum is created:
specifically in testing, when a task ends with status dep_fail.

In such a case, data is not stored properly to the monitoring database.
This failure is logged in database_manager.log but is not made
obvious to the user/CI.