Fix race condition with wait_for_tasks vs task completion
Prior to this commit, if wait_for_tasks was waiting for task completion,
and some tasks completed while it was waiting, then

> RuntimeError: dictionary changed size during iteration

would be raised due to the task dictionary being modified.

This commit duplicates the list of task dictionary keys at the start of
iteration, and so the task dictionary can subsequently be
modified without causing this error.

Fixes #1606