Pin azure dependency
The azure dependency broke when it moved to version 5.0.0. This
commit pins azure to major version 4, which is the major version
installed in my dev environment at the moment.

That breakage was blocking CI. This commit makes a minimal change
to unblock CI.

If azure is going to be used a lot, it would make sense to test
against azure version >=5, and move to correct dependencies with
a lower bound rather than an upper bound on older dependencies.