Test that parsl-visualize does not return HTTP errors
This wget recursively follows every link starting at the top level page of
a parsl-visualize web server, and checks that each of them returns with
a successful response, rather than a server error.

This commit also fixes some pages so that this test can pass, by replacing
some pages which returned a redirect (to add /) with the target of that redirect.